STABLE/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib==2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.15.4
motor==3.7.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests==2.32.5
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
sendgrid>=6.11.0
httpx==0.28.1
coinbase-advanced-py==1.8.2
cbpro
cdp-sdk
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.50.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
twilio==9.8.7
aiohttp
user-agents==2.2.0
python-telegram-bot==21.0
pyotp==2.9.0
qrcode==8.2
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0
redis==5.0.1
aioredis==2.0.1
fastapi-cache2[redis]==0.2.1
slowapi==0.1.9

